requirements.txt
aiohttp==3.8.4
aioredis==2.0.1
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.2
attrs==23.1.0
bcrypt==4.0.1
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.4
colorama==0.4.6
dnspython==2.3.0
email-validator==2.0.0.post2
fastapi==0.100.0
fastapi-cache2==0.2.1
frozenlist==1.4.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
passlib==1.7.4
pendulum==2.1.2
pydantic==2.0.2
pydantic_core==2.1.2
PyJWT==2.7.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytzdata==2020.1
redis==4.6.0
requests==2.31.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.18
starlette==0.27.0
typing_extensions==4.7.1
urllib3==2.0.3
uvicorn==0.22.0
yarl==1.9.2
